all: Move EntityKey to graph::schema
