+++ graph/src/schema/mod.rs
@@ -27,6 +27,7 @@ mod api;
 /// Utilities for working with GraphQL schema ASTs.
 pub mod ast;
 
+mod entity_key;
 mod entity_type;
 mod fulltext;
 mod input_schema;
@@ -34,6 +35,7 @@ mod input_schema;
 pub use api::{api_schema, is_introspection_field, APISchemaError, INTROSPECTION_QUERY_TYPE};
 
 pub use api::{ApiSchema, ErrorPolicy};
+pub use entity_key::EntityKey;
 pub use entity_type::{AsEntityTypeName, EntityType};
 pub use fulltext::{FulltextAlgorithm, FulltextConfig, FulltextDefinition, FulltextLanguage};
 pub use input_schema::InputSchema;
